+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to
 
 ### Changed
 * Changed `EthernetUDP::localPort()` to be `const`.
+* Changed `entropy_random_range()` to return zero if `EAGAIN`.
 
 ### Fixed
 * Fixed `SNTP_SET_SYSTEM_TIME_US(sec, us)` definition to set the RTC directly